python/requirements.txt
deltadefi==0.0.4
python-dotenv>=1.0.0
sidan_gin==0.1.1